Support uploading  project files on watch
