+++ packages/cli-lib/lang/en.lyaml
@@ -481,13 +481,14 @@ en:
             describe: "Watch your local project for changes and automatically upload changed files to a new build in HubSpot"
             examples:
               default: "Watch a project within the myProjectFolder folder"
-            logs: 
+            logs:
               watching: "Watcher is ready and watching \"{{ projectDir }}\". Any changes detected will be automatically uploaded and added to the current staging build."
               resuming: "Resuming watcher..."
               uploadSucceeded: "Uploaded file \"{{ filePath }}\" to \"{{ remotePath }}\""
               deleteSucceeded: "Deleted file \"{{ remotePath }}\""
               createNewBuild: "Staging build #{{ buildId }} created"
               buildCancelled: "Staging build has been cancelled. Please try running `hs project watch` again"
+              initialUploadRequired: "This project has no builds. Uploading files."
             debug:
               attemptNewBuild: "Attempting to create a new build"
               uploading: "Attempting to upload file \"{{ filePath }}\" to \"{{ remotePath }}\""
@@ -501,7 +502,7 @@ en:
               projectLocked: "The previous staging build is still in progress"
               uploadFailed: "Failed to upload file \"{{ filePath }}\" to \"{{ remotePath }}\""
               deleteFailed: "Failed to delete file \"{{ remotePath }}\""
-      remove: 
+      remove:
         describe: "Delete a file or folder from HubSpot."
         deleted: "Deleted \"{{ path }}\" from account {{ accountId }}"
         errors:
