+++ packages/cli/lib/projects.js
@@ -1,6 +1,7 @@
 const fs = require('fs-extra');
 const path = require('path');
-
+const archiver = require('archiver');
+const tmp = require('tmp');
 const chalk = require('chalk');
 const findup = require('findup-sync');
 const Spinnies = require('spinnies');
@@ -22,15 +23,18 @@ const {
   getBuildStatus,
   getDeployStatus,
   fetchProject,
+  uploadProject,
 } = require('@hubspot/cli-lib/api/dfs');
 const {
   logApiErrorInstance,
   ApiErrorContext,
 } = require('@hubspot/cli-lib/errorHandlers');
+const { shouldIgnoreFile } = require('@hubspot/cli-lib/ignoreRules');
 const { getCwd } = require('@hubspot/cli-lib/path');
 const { promptUser } = require('./prompts/promptUtils');
 const { EXIT_CODES } = require('./enums/exitCodes');
 const { uiLine, uiAccountDescription } = require('../lib/ui');
+const { i18n } = require('@hubspot/cli-lib/lib/lang');
 
 const PROJECT_STRINGS = {
   BUILD: {
@@ -232,6 +236,108 @@ const getProjectDetailUrl = (projectName, accountId) => {
   return `${baseUrl}/developer-projects/${accountId}/project/${projectName}`;
 };
 
+const uploadProjectFiles = async (accountId, projectName, filePath) => {
+  const i18nKey = 'cli.commands.project.subcommands.upload';
+  const spinnies = new Spinnies({
+    succeedColor: 'white',
+  });
+  const accountIdentifier = uiAccountDescription(accountId);
+
+  spinnies.add('upload', {
+    text: i18n(`${i18nKey}.loading.upload.add`, {
+      accountIdentifier,
+      projectName,
+    }),
+  });
+
+  let buildId;
+
+  try {
+    const upload = await uploadProject(accountId, projectName, filePath);
+
+    buildId = upload.buildId;
+
+    spinnies.succeed('upload', {
+      text: i18n(`${i18nKey}.loading.upload.succeed`, {
+        accountIdentifier,
+        projectName,
+      }),
+    });
+
+    logger.debug(
+      i18n(`${i18nKey}.debug.buildCreated`, {
+        buildId,
+        projectName,
+      })
+    );
+  } catch (err) {
+    spinnies.fail('upload', {
+      text: i18n(`${i18nKey}.loading.upload.fail`, {
+        accountIdentifier,
+        projectName,
+      }),
+    });
+
+    logApiErrorInstance(
+      err,
+      new ApiErrorContext({
+        accountId,
+        projectName,
+      })
+    );
+    process.exit(EXIT_CODES.ERROR);
+  }
+
+  return { buildId };
+};
+
+const handleProjectUpload = async (
+  accountId,
+  projectConfig,
+  projectDir,
+  callbackFunc
+) => {
+  const i18nKey = 'cli.commands.project.subcommands.upload';
+  const tempFile = tmp.fileSync({ postfix: '.zip' });
+
+  logger.debug(
+    i18n(`${i18nKey}.debug.compressing`, {
+      path: tempFile.name,
+    })
+  );
+
+  const output = fs.createWriteStream(tempFile.name);
+  const archive = archiver('zip');
+
+  output.on('close', async function() {
+    logger.debug(
+      i18n(`${i18nKey}.debug.compressed`, {
+        byteCount: archive.pointer(),
+      })
+    );
+
+    const { buildId } = await uploadProjectFiles(
+      accountId,
+      projectConfig.name,
+      tempFile.name
+    );
+
+    if (callbackFunc) {
+      callbackFunc(tempFile, buildId);
+    }
+  });
+
+  archive.pipe(output);
+
+  archive.directory(
+    path.resolve(projectDir, projectConfig.srcDir),
+    false,
+    file => (shouldIgnoreFile(file.name) ? false : file)
+  );
+
+  archive.finalize();
+};
+
 const showWelcomeMessage = () => {
   logger.log('');
   logger.log(chalk.bold('Welcome to HubSpot Developer Projects!'));
@@ -392,6 +498,7 @@ module.exports = {
   writeProjectConfig,
   getProjectConfig,
   getIsInProject,
+  handleProjectUpload,
   createProjectConfig,
   validateProjectConfig,
   showWelcomeMessage,
